@@ -16,7 +16,7 @@ import Foundation
 /// - nationalNumber: National number as an unsigned. Int.
 /// - numberExtension: Extension if available. String. Optional
 /// - type: Computed phone number type on access. Returns from an enumeration - PNPhoneNumberType.
-public struct PhoneNumber {
+public struct PhoneNumber: Sendable {
     public let numberString: String
     public let countryCode: UInt64
     public let leadingZero: Bool
